Move hints and related errors to a separate module
