+++ modules/generic-base/src/main/scala/pureconfig/generic/ProductHint.scala
